Privacy respecting approach to counting installed versions
When the version is updated, create a random number.

Send this random number to the backend server exactly once when checking if a
new version is available, then reset it to 0. This way the backend can count
the number of different random numbers it sees for a specific version, but
since this number is immediately discarded afterwards, it cannot be used to
track a user.

Note that the backend doesn't keep log files on accesses to the `/status` api.

Yes, this metric can trivially be gamed if someone wants to mess with our
attempts to get an idea of the number of users. Let's assume people aren't
jerks about this.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>