docker networking: avoid address blocks used by local networks
Grab the networks of all physical interfaces and check a large set of candidate
cidrs in order to find one to use that won't conflict with an address range
that's already in use.

Add a pre execution script to adsb-docker that sets that available cidr as the
base for Docker's network bridge creation.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>