add USB TEMPer style sensor support
This appears to support quite a few USB temperature sensors. Which should
work on most platforms.

The file temper.py comes from https://github.com/ccwienk/temper where it
is provided under the MIT license and can therefore be incorporated here.

The version here has a couple of tiny modifications to adapt it to the use
in this project.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>